+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Evnt changelog
 
-## Version 3.0.2 IN-PROGRESS
+## Version 3.0.2
 
 - Added initializer generator for Ruby on Rails application.
 - Added command generator for Ruby on Rails application.
 - Added event generator for Ruby on Rails application.
 - Added handler generator for Ruby on Rails application.
 - Fixed required parameters on command and event initializer.
+- Fixed command exception error.
 
 ## Version 3.0.1